examples updated
