+++ analytics/terraform/spark-k8s-operator/examples/benchmark/tpcds-benchmark-1t.yaml
@@ -1,17 +1,14 @@
 # NOTE: This example requires the following prerequisites before executing the jobs
 # 1. Ensure spark-team-a name space exists
-# 2. replace <ENTER_YOUR_BUCKET> with your bucket name
+# 2. replace <ENTER_YOUR_BUCKET_NAME> with your bucket name
 # 3. Ensure you run "analytics/spark-k8s-operator/spark-samples/tpcds-benchmark-data-generation-1t.yaml" to generate the INPUT data in S3 bucket and update INPUT argument(      "s3a://<ENTER_YOUR_BUCKET>/TPCDS-TEST-1T/catalog_sales/") path in the below yaml
 
 ---
 apiVersion: "sparkoperator.k8s.io/v1beta2"
 kind: SparkApplication
 metadata:
-  name: tpcds-benchmark-10g
+  name: tpcds-benchmark-1t
   namespace: spark-team-a
-  labels:
-    yunikorn.apache.org/username: "vara"
-    queue: "root.spark-team-a"
 spec:
   type: Scala
   pythonVersion: "3"
@@ -21,15 +18,6 @@ spec:
   imagePullPolicy: Always
   mainClass: com.amazonaws.eks.tpcds.BenchmarkSQL
   mainApplicationFile: local:///opt/spark/examples/jars/eks-spark-benchmark-assembly-1.0.jar
-  #  volumes:
-  #    - name: "spark-local-dir-1"     # Support mapping spark.local.dir to hostPath volume
-  #      hostPath:
-  #        path: "/tmp/mnt-1"
-  batchScheduler: yunikorn
-  #  batchSchedulerOptions:
-  #    resources:
-  #      cpu: "2"
-  #      memory: "4096m"
   arguments:
     # TPC-DS data location
     - "s3a://<ENTER_YOUR_BUCKET_NAME>/TPCDS-TEST-1T"
@@ -72,11 +60,7 @@ spec:
     "spark.hadoop.fs.s3a.input.fadvise": "random"
     "spark.hadoop.fs.s3a.impl": "org.apache.hadoop.fs.s3a.S3AFileSystem"
     "spark.sql.execution.arrow.pyspark.enabled": "true"
-    # YuniKorn Config
-    "spark.kubernetes.driver.label.queue": "root.spark-team-a"
-    "spark.kubernetes.driver.label.yunikorn.apache.org/username": "vara"
-    "spark.kubernetes.executor.label.queue": "root.spark-team-a"
-    "spark.kubernetes.executor.label.yunikorn.apache.org/username": "vara"
+
   restartPolicy:
     type: OnFailure
     onFailureRetries: 3
@@ -86,18 +70,19 @@ spec:
   volumes: # using NVMe instance storage mounted on /local1
     - name: spark-local-dir-1
       hostPath:
-        path: /local1/benchmark
-        type: DirectoryOrCreate
+        path: /local1
+        type: Directory
   driver:
-    volumeMounts: # Points to InstanceStore 150GB NVMe SSD for shuffle spill over from memory
+    volumeMounts: # Points to InstanceStore NVMe SSD for shuffle spill over from memory
       - name: spark-local-dir-1
-        mountPath: /local1/benchmark
+        mountPath: /data1
+        readOnly: false
     initContainers:
       - name: volume-permissions
         image: public.ecr.aws/y4g4v0z7/busybox
-        command: [ 'sh', '-c', 'chown -R 185 /local1/benchmark' ]
+        command: [ 'sh', '-c', 'chown -R 185 /local1' ]
         volumeMounts:
-          - mountPath: "/local1/benchmark"
+          - mountPath: "/local1"
             name: "spark-local-dir-1"
     cores: 1
     memory: "10g"
@@ -106,18 +91,16 @@ spec:
     labels:
       version: 3.1.1
     nodeSelector:
-      NodeGroupType: "spark"
+      NodeGroupType: "SparkMemoryOptimized"
+    tolerations:
+      - key: "spark-memory-optimized"
+        operator: "Exists"
+        effect: "NoSchedule"
   executor:
     volumeMounts:
       - name: spark-local-dir-1
-        mountPath: /local1/benchmark
-    initContainers:
-      - name: volume-permissions
-        image: public.ecr.aws/y4g4v0z7/busybox
-        command: [ 'sh', '-c', 'chown -R 185 /local1/benchmark' ]
-        volumeMounts:
-          - mountPath: "/local1/benchmark"
-            name: "spark-local-dir-1"
+        mountPath: /data1
+        readOnly: false
     cores: 1
     instances: 10
     memory: "6g"
@@ -126,4 +109,8 @@ spec:
     labels:
       version: 3.1.1
     nodeSelector:
-      NodeGroupType: "spark"
+      NodeGroupType: "SparkMemoryOptimized"
+    tolerations:
+      - key: "spark-memory-optimized"
+        operator: "Exists"
+        effect: "NoSchedule"
