+++ analytics/terraform/spark-k8s-operator/examples/cluster-autoscaler/pyspark-pi-job.yaml
@@ -1,0 +1,49 @@
+# NOTE: This example requires the following prerequisites before executing the jobs
+# 1. Ensure spark-team-a name space exists
+
+---
+apiVersion: "sparkoperator.k8s.io/v1beta2"
+kind: SparkApplication
+metadata:
+  name: pyspark-pi
+  namespace: spark-team-a
+spec:
+  type: Python
+  pythonVersion: "3"
+  mode: cluster
+  image: "public.ecr.aws/r1l5w1y9/spark-operator:3.2.1-hadoop-3.3.1-java-11-scala-2.12-python-3.8-latest"
+  imagePullPolicy: Always
+  mainApplicationFile: local:///opt/spark/examples/src/main/python/pi.py
+  sparkVersion: "3.1.1"
+  restartPolicy:
+    type: OnFailure
+    onFailureRetries: 3
+    onFailureRetryInterval: 10
+    onSubmissionFailureRetries: 5
+    onSubmissionFailureRetryInterval: 20
+  driver:
+    cores: 1
+    coreLimit: "1200m"
+    memory: "512m"
+    labels:
+      version: 3.1.1
+    serviceAccount: spark-team-a
+    nodeSelector:
+      NodeGroupType: "spark-on-demand-ca"
+    tolerations:
+      - key: "spark-on-demand-ca"
+        operator: "Exists"
+        effect: "NoSchedule"
+  executor:
+    cores: 1
+    instances: 2
+    memory: "512m"
+    serviceAccount: spark-team-a
+    labels:
+      version: 3.1.1
+    nodeSelector:
+      NodeGroupType: "spark-spot-ca"
+    tolerations:
+      - key: "spark-spot-ca"
+        operator: "Exists"
+        effect: "NoSchedule"
