--- analytics/terraform/spark-k8s-operator/examples/nvme-ephemeral-storage/nvme-ephemeral-storage.yaml
@@ -1,114 +1,0 @@
-# Pre-requisite before running this job
-# 1/ Open taxi-trip-execute.sh and update <S3_BUCKET> , <JOB_NAME>  and <REGION>
-# 2/ Replace <S3_BUCKET>  and <JOB_NAME> with your S3 bucket in this yaml
-# 3/ execute taxi-trip-execute.sh
-
----
-apiVersion: "sparkoperator.k8s.io/v1beta2"
-kind: SparkApplication
-metadata:
-  name: "<JOB_NAME>"
-  namespace: spark-team-a
-  labels:
-    app: "<JOB_NAME>"
-    applicationId: "<JOB_NAME>-yunikorn"
-    queue: root.test
-spec:
-  type: Python
-  sparkVersion: "3.2.1"
-  pythonVersion: "3"
-  mode: cluster
-  # Public docker image used from datamechanics/spark:3.2.1-hadoop-3.3.1-java-11-scala-2.12-python-3.8-latest and pushed to ECR repo
-  image: "public.ecr.aws/r1l5w1y9/spark-operator:3.2.1-hadoop-3.3.1-java-11-scala-2.12-python-3.8-latest"
-  imagePullPolicy: IfNotPresent
-  mainApplicationFile: "s3a://<S3_BUCKET>/<JOB_NAME>/scripts/pyspark-taxi-trip.py"  # MainFile is the path to a bundled JAR, Python, or R file of the application
-  arguments:
-    - "s3a://<S3_BUCKET>/<JOB_NAME>/input/"
-    - "s3a://<S3_BUCKET>/<JOB_NAME>/output/"
-  hadoopConf:
-    "fs.s3a.aws.credentials.provider": "com.amazonaws.auth.WebIdentityTokenCredentialsProvider"
-    "fs.s3a.impl": "org.apache.hadoop.fs.s3a.S3AFileSystem"
-    "mapreduce.fileoutputcommitter.algorithm.version": "2"
-  sparkConf:
-    "spark.speculation": "false"
-    "spark.network.timeout": "2400"
-    "spark.hadoop.fs.s3a.connection.timeout": "1200000"
-    "spark.hadoop.fs.s3a.path.style.access": "true"
-    "spark.hadoop.fs.s3a.connection.maximum": "200"
-    "spark.hadoop.fs.s3a.fast.upload": "true"
-    "spark.hadoop.fs.s3a.readahead.range": "256K"
-    "spark.hadoop.fs.s3a.input.fadvise": "random"
-    "spark.hadoop.fs.s3a.impl": "org.apache.hadoop.fs.s3a.S3AFileSystem"
-
-    #Event logs
-    "spark.local.dir": "/data1"
-    "spark.eventLog.enabled": "true"
-    "spark.eventLog.dir": "s3a://<ENTER_S3_BUCKET>/logs"
-
-    # Expose Spark metrics for Prometheus
-    "spark.ui.prometheus.enabled": "true"
-    "spark.executor.processTreeMetrics.enabled": "true"
-    "spark.kubernetes.driver.annotation.prometheus.io/scrape": "true"
-    "spark.kubernetes.driver.annotation.prometheus.io/path": "/metrics/executors/prometheus/"
-    "spark.kubernetes.driver.annotation.prometheus.io/port": "4040"
-    "spark.kubernetes.driver.service.annotation.prometheus.io/scrape": "true"
-    "spark.kubernetes.driver.service.annotation.prometheus.io/path": "/metrics/driver/prometheus/"
-    "spark.kubernetes.driver.service.annotation.prometheus.io/port": "4040"
-    "spark.metrics.conf.*.sink.prometheusServlet.class": "org.apache.spark.metrics.sink.PrometheusServlet"
-    "spark.metrics.conf.*.sink.prometheusServlet.path": "/metrics/driver/prometheus/"
-    "spark.metrics.conf.master.sink.prometheusServlet.path": "/metrics/master/prometheus/"
-    "spark.metrics.conf.applications.sink.prometheusServlet.path": "/metrics/applications/prometheus/"
-
-  restartPolicy:
-    type: OnFailure
-    onFailureRetries: 3
-    onFailureRetryInterval: 10
-    onSubmissionFailureRetries: 5
-    onSubmissionFailureRetryInterval: 20
-  volumes:  # using NVMe instance storage mounted on /local1
-    - name: spark-local-dir-1
-      hostPath:
-        path: /local1
-        type: Directory
-
-  driver:
-    volumeMounts: # Points to InstanceStore 150GB NVMe SSD for shuffle spill over from memory
-      - name: spark-local-dir-1
-        mountPath: /data1
-        readOnly: false
-    initContainers:
-      - name: volume-permissions
-        image: public.ecr.aws/y4g4v0z7/busybox
-        command: [ 'sh', '-c', 'chown -R 185 /local1' ]
-        volumeMounts:
-          - mountPath: "/local1"
-            name: "spark-local-dir-1"
-    cores: 1
-    memory: "4g"
-    serviceAccount: spark-team-a
-    labels:
-      version: 3.2.1
-    nodeSelector:
-      NodeGroupType: "spark"
-  executor:
-    podSecurityContext:
-      fsGroup: 185
-    volumeMounts:
-      - name: spark-local-dir-1
-        mountPath: /data1
-        readOnly: false
-    initContainers:
-      - name: volume-permissions
-        image: public.ecr.aws/y4g4v0z7/busybox
-        command: [ 'sh', '-c', 'chown -R 185 /local1' ]
-        volumeMounts:
-          - mountPath: "/local1"
-            name: "spark-local-dir-1"
-    cores: 1
-    instances: 4
-    memory: "4g"
-    serviceAccount: spark-team-a
-    labels:
-      version: 3.2.1
-    nodeSelector:
-      NodeGroupType: "spark"
