+++ analytics/terraform/spark-k8s-operator/helm-values/spark-operator-values.yaml
@@ -6,6 +6,12 @@ webhook:
   # -- Webhook service port
   port: 8080
 
+# -- Set this if running spark jobs in a different namespace than the operator
+sparkJobNamespace: "spark-team-a"
+
+# -- Operator concurrency, higher values might increase memory usage
+controllerThreads: 10
+
 # resources -- Pod resource requests and limits
 # Note, that each job submission will spawn a JVM within the Spark Operator Pod using "/usr/local/openjdk-11/bin/java -Xmx128m".
 # Kubernetes may kill these Java processes at will to enforce resource limits. When that happens, you will see the following error:
